Refactored back button
Moved back button from CustomNavbar to App component. This makes it more clear where it should be in the code, since that's how it appears on the page. I also fixed the positioning to make the button absolute. This allows it to not affect the spacing of the rest of the pages it's on.